bug 2512: fixed whoops in complex double
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@26971 0c594bb0-8543-0410-a25e-d8ec440261db

+++ IdlMatlabSupport/src/org/autoplot/idlsupport/ReadIDLSav.java
@@ -329,7 +329,7 @@ public class ReadIDLSav {
                 case TYPECODE_COMPLEX_FLOAT:
                     return new float[] { buf.getFloat(offs), buf.getFloat(offs+4) };
                 case TYPECODE_COMPLEX_DOUBLE:
-                    return new double[] { buf.getDouble(offs), buf.getFloat(offs+8) };
+                    return new double[] { buf.getDouble(offs), buf.getDouble(offs+8) };
                 case TYPECODE_STRING:
                     int len= buf.getInt(offs);
                     if ( len<0 || len>1024 ) {
